# docs/swagger-old.yaml
openapi: 3.0.0
info:
  title: APIServiceGo
  version: 1.0.0
  description: EV Charging API for searching charging connectors, getting estimates, managing orders, and processing payments for electric vehicle charging sessions.
servers:
  - url: http://{{localhost}}
paths:
...
  /v1/orders/{order_id}:
    get:
      summary: get order info
      tags:
        - Orders
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Bpp-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                default: <string>
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
              example:
                order:
                  id: order-bpp-789012
                  status: ACTIVE
                  mode: RESERVATION
                payment:
                  status: PAID
                charging:
                  status: ACTIVE
                connectorId: ""
                connectorType: ""
                vehicle:
                  type: ""
                  model: ""
                  make: ""
                trackingUrl: https://track.bluechargenet-aggregator.io/session/SESSION-9876543210
                chargingTelemetry:
                  eventTime: "2025-01-27T17:00:00Z"
                  metrics:
                    - name: STATE_OF_CHARGE
                      value: 62.5
                      unitCode: PERCENTAGE
                    - name: POWER
                      value: 18.4
                      unitCode: KWH
                    - name: ENERGY
                      value: 10.2
                      unitCode: KW
                    - name: VOLTAGE
                      value: 392
                      unitCode: VLT
                    - name: CURRENT
                      value: 47
                      unitCode: AMP
                    - name: SESSION_DURATION
                      value: 10
                      unitCode: min
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - Order not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Order not found."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: >-
        Retrieve order details, current status, and telemetry for the specified
        order.
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_REQUEST"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
    Vehicle:
      type: object
      description: Vehicle information
      properties:
        make:
          type: string
          description: Vehicle manufacturer
        model:
          type: string
          description: Vehicle model
        type:
          type: string
          enum:
            - 2-wheeler
            - 3-wheeler
            - 4-wheeler
          description: Vehicle type
...
    OrderInfo:
      type: object
      description: Order information
      properties:
        id:
          type: string
          description: Order identifier
        mode:
          type: string
          description: Order mode (e.g., reservation, RESERVATION)
        status:
          type: string
          description: Order status (e.g., quoted_price, ACTIVE, CANCELLED, COMPLETED)
    PaymentInfo:
      type: object
      description: Payment information
      properties:
        status:
          type: string
          description: Payment status (e.g., PAID, PAID)
    ChargingInfo:
      type: object
      description: Charging session information
      properties:
        status:
          type: string
          description: Charging status (e.g., ACTIVE, COMPLETED)
...
    ChargingMetric:
      type: object
      description: Charging telemetry metric
      properties:
        name:
          type: string
          description: Metric name (e.g., STATE_OF_CHARGE, POWER, ENERGY, VOLTAGE, CURRENT, SESSION_DURATION)
        value:
          type: number
          format: float
          description: Metric value
        unitCode:
          type: string
          description: Unit code for the metric (e.g., PERCENTAGE, KWH, KW, VLT, AMP, min)
    ChargingTelemetry:
      type: object
      description: Real-time charging telemetry data
      properties:
        eventTime:
          type: string
          format: date-time
          description: Timestamp of the telemetry event
        metrics:
          type: array
          description: Array of charging metrics
          items:
            $ref: "#/components/schemas/ChargingMetric"
    OrderResponse:
      type: object
      description: Order response with details
      properties:
        order:
          $ref: "#/components/schemas/OrderInfo"
        payment:
          $ref: "#/components/schemas/PaymentInfo"
        charging:
          $ref: "#/components/schemas/ChargingInfo"
        connectorId:
          type: string
          description: Connector identifier
        connectorType:
          type: string
          description: Type of connector
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        trackingUrl:
          type: string
          format: uri
          description: URL to track the charging session
        chargingTelemetry:
          $ref: "#/components/schemas/ChargingTelemetry"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []
...